Separate timelineHeader from timeline
Improve houx library

Change composing reducer logic from loop to reduce

Remove types

Improve styles

Remove explicit any type

Remove setup and change navigation actions
