+++ docs/src/modules/redux/features/navigation/actionTypes.ts
@@ -1,3 +1,1 @@
-export const SETUP_NAVIGATION = 'SETUP_NAVIGATION';
-export const CHANGE_NAVIGATION = 'CHANGE_NAVIGATION';
 export const LOAD_PAGES = 'LOAD_PAGES';
