+++ docs/src/modules/redux/features/navigation/reducer.ts
@@ -2,7 +2,7 @@ import { Page } from '../../../../../../src/typings/data/import';
 import { loadPages } from '../../../../pages';
 import { determineActivePage, flattenPages } from '../../../utils/router';
 import { StoreAction } from '../actionType';
-import { CHANGE_NAVIGATION, LOAD_PAGES, SETUP_NAVIGATION } from './actionTypes';
+import { LOAD_PAGES } from './actionTypes';
 
 interface Props {
   pages: Array<Page>;
@@ -21,16 +21,6 @@ export const initialState: Props = {
 
 const navigationReducer = (state = initialState, action: StoreAction) => {
   switch (action.type) {
-    case SETUP_NAVIGATION:
-      return {
-        ...state,
-        pages: action.payload.pages
-      };
-    case CHANGE_NAVIGATION:
-      return {
-        ...state,
-        activePage: action.payload.activePage
-      };
     case LOAD_PAGES: {
       const pages = loadPages(action.payload.pathname, state.pages);
       const flattenedPages = flattenPages(pages);
