--- docs/src/modules/redux/features/navigation/type.ts
@@ -1,9 +1,0 @@
-export interface Page {
-  pathname: string;
-  title?: string;
-  children?: Array<Page>;
-  displayNav?: boolean;
-  subheader?: string;
-  icon?: string;
-  highlight?: boolean;
-}
