+++ src/components/device/browser/views/HeaderView.tsx
@@ -1,9 +1,9 @@
-import { createStyles, makeStyles } from '@material-ui/core';
+import { createStyles, makeStyles, Theme } from '@material-ui/core';
 import React, { FC } from 'react';
 
 import { ChromeInput } from '../components/ChromeInput';
 
-const useStyles = makeStyles(() =>
+const useStyles = makeStyles((theme: Theme) =>
   createStyles({
     browserBar: {
       display: 'flex',
@@ -11,14 +11,14 @@ const useStyles = makeStyles(() =>
     },
     circleContainer: {
       display: 'flex',
-      padding: '8px'
+      padding: theme.spacing(1)
     },
     circle: {
       backgroundColor: '#b6c1cd',
-      width: '10px',
-      height: '10px',
-      borderRadius: '5px',
-      marginLeft: '5px'
+      width: '12px',
+      height: '12px',
+      borderRadius: '6px',
+      marginLeft: '6px'
     }
   })
 );
